Move apache_util.py to _internal
