+++ certbot-apache/certbot_apache/override_centos.py
@@ -11,7 +11,7 @@ from certbot.errors import MisconfigurationError
 
 from acme.magic_typing import List  # pylint: disable=unused-import, no-name-in-module
 
-from certbot_apache import apache_util
+from certbot_apache._internal import apache_util
 from certbot_apache import configurator
 from certbot_apache import parser
