+++ certbot-apache/certbot_apache/override_fedora.py
@@ -6,7 +6,7 @@ from certbot import errors
 from certbot import interfaces
 from certbot import util
 
-from certbot_apache import apache_util
+from certbot_apache._internal import apache_util
 from certbot_apache import configurator
 from certbot_apache import parser
